Removing tkrtypes directory by moving in required code to tkg/registry directory
