+++ tkg/registry/util_test.go
@@ -1,7 +1,7 @@
 // Copyright 2021 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package tkrtypes
+package registry
 
 import (
 	"fmt"
