--- tkg/tkrtypes/doc.go
@@ -1,6 +1,0 @@
-// Copyright 2021 VMware, Inc. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// Package tkrtypes defines some common types used in TKR.
-// Copied over from github.com/vmware-tanzu/tanzu-framework/pkg/v1/tkr/pkg/types
-package tkrtypes
